+++ CHANGELOG.md
@@ -13,9 +13,8 @@
 * Added function `Attachment.getFilename`. This function is used to get the name an attachment will be saved with given the specified arguments. Arguments are identical to `Attachment.save`.
 * Added zip file support for the `Attachment.save` and `Message.save`. Simply pass a path for the `zip` keyword argument and it will create a new ZipFile instance and save all of it's data inside there. Alternatively, you can pass an instance of a class that is either a ZipFile or ZipFile-like and it will simply use that. When this argument is defined, the `customPath` argument refers to the path inside the zip file.
 * Added the `html` and `rtf` keywords to `Message.save`. These will attempt to save the body in the html or rtf format, respectively. If the program cannot save in those formats, it will raise an exception unless the `allowFallback` keyword argument is `True`.
+* Changed `utils.hasLen` to use `hasattr` instead of the try-except method it was using.
 * Many renames to better fit naming convention:
-    * `utils.get_input` to `utils.getInput`.
-    * `utils.setup_logging` to `utils.setupLogging`.
     * `dev.setup_dev_logger` to `dev.setupDevLogger`.
     * `MSGFile.fix_path` to `MSGFile.fixPath`.
     * `MessageBase.save_attachments` to `MessageBase.saveAttachments`.
@@ -27,9 +26,14 @@
     * `Properties.next_recipient_id` to `Properties.nextRecipientId`.
     * `Properties.recipient_count` to `Properties.recipientCount`.
     * `utils.get_command_args` to `utils.getCommandArgs`.
+    * `utils.get_full_class_name` to `utils.getFullClassName`.
+    * `utils.get_input` to `utils.getInput`.
+    * `utils.has_len` to `utils.hasLen`.
+    * `utils.setup_logging` to `utils.setupLogging`.
     * `constants.int_to_data_type` to `constants.intToDataType`.
     * `constants.int_to_intelligence` to `constants.intToIntelligence`.
     * `constants.int_to_recipient_type` to `constants.intToRecipientType`.
+    * Misc internal function variables.
 
 **v0.28.7**
 * Added hex versions of the `MULTIPLE_X_BYTES` constants.